+++ CHANGELOG.md
@@ -8,6 +8,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 [Please read through the Keep a Changelog (~5min)](https://keepachangelog.com/en/1.0.0/).
 
+## [Unreleased] - YYYY-MM-DD
+
+## Added
+
+- `--jinja-variables-file` for `dbx execute`
 
 ----
 > Unreleased changes must be tracked above this line.